+++ lib/src/peppermint_utility.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:file_picker/file_picker.dart';
@@ -7,9 +8,9 @@ import 'package:peppermint_sdk/peppermint_sdk.dart';
 import 'package:peppermint_sdk/src/peppermint_constants.dart';
 import 'package:peppermint_sdk/src/widgets/image_crop_view.dart';
 import 'package:scan/scan.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'widgets/camera_view.dart';
 import 'widgets/photo_filters/image_editor.dart';
-import 'widgets/scanner_view.dart';
 
 class PeppermintUtility {
   PeppermintUtility._();
@@ -42,39 +43,46 @@ class PeppermintUtility {
   }
 
   /// Open file explorer and select media
-  static Future<File?> getMediaFromExplorer() async {
+  static Future<File?> getMediaFromExplorer({bool squareCrop = true}) async {
     FilePickerResult? result = await FilePicker.platform.pickFiles(
       type: FileType.custom,
       allowedExtensions: PeppermintConstants.fileTypes,
     );
+    File? file;
     if (result != null) {
       String fileType = PeppermintUtility.getFileType(
           path.extension(result.files.single.path!));
       if (fileType == 'image') {
         /// Crop square image
-
-        File? cropped =
-            await Get.to(() => ImageCropView(File(result.files.single.path!)));
-        return cropped;
-      } else {
-        return File(result.files.single.path!);
+        if (squareCrop) {
+          File? cropped = await Get.to(
+              () => ImageCropView(File(result.files.single.path!)));
+          file = cropped;
+        } else {
+          file = File(result.files.single.path!);
+        }
       }
     }
-    return null;
+    return file;
   }
 
   /// Get image from gallery
-  static Future<File?> getImageFromGallery() async {
+  static Future<File?> getImageFromGallery({bool squareCrop = true}) async {
     FilePickerResult? result = await FilePicker.platform.pickFiles(
       type: FileType.image,
     );
-    if (result != null) {
-      File? cropped =
-          await Get.to(() => ImageCropView(File(result.files.single.path!)));
+    File? file;
 
-      return cropped;
+    if (result != null) {
+      if (squareCrop) {
+        File? cropped =
+            await Get.to(() => ImageCropView(File(result.files.single.path!)));
+        file = cropped;
+      } else {
+        file = File(result.files.single.path!);
+      }
     }
-    return null;
+    return file;
   }
 
   /// Get video from gallery
@@ -89,23 +97,25 @@ class PeppermintUtility {
   }
 
   /// Get image from camera, it is cropped square and go to image editor first
-  static Future<File?> getImageFromCamera() async {
-    File? image;
+  static Future<File?> getImageFromCamera({bool edit = true}) async {
     File? file = await Get.to(() => const CameraView());
     if (file != null) {
-      Uint8List imageData = file.readAsBytesSync();
-      image = await Get.to(
-        () => ImageEditor(
-          image: imageData,
-          savePath: file.path,
-          allowCamera: true,
-          allowGallery: true,
-        ),
-      );
+      if (edit) {
+        Uint8List imageData = file.readAsBytesSync();
+        file = await Get.to(
+          () => ImageEditor(
+            image: imageData,
+            savePath: file!.path,
+            allowCamera: true,
+            allowGallery: true,
+          ),
+        );
+      }
     }
-    return image;
+    return file;
   }
 
+  /// Scan QR from image uploaded.
   static Future<QRResult> getQRFile() async {
     FilePickerResult? result = await FilePicker.platform.pickFiles(
       type: FileType.image,
@@ -122,8 +132,36 @@ class PeppermintUtility {
     return QRResult(success: true);
   }
 
-  static Future<String?> scanQR() async {
-    String? result = await Get.to(() => const ScannerView());
-    return result;
+  /// launch url via browser.
+  static launchBrowser(String url) async {
+    bool success = true;
+    if (!url.contains('http')) {
+      url = 'http://$url';
+    }
+    if (url.contains('facebook.com')) {
+      url = 'fb://facewebmodal/f?href=$url';
+    }
+    if (await canLaunchUrl(Uri.parse(url))) {
+      if (!await launchUrl(
+        Uri.parse(url),
+        mode: LaunchMode.externalApplication,
+      )) {
+        success = false;
+      }
+    } else {
+      success = false;
+    }
+    return success;
+  }
+
+  /// Generate random name for new contract.
+  static String generateContractName() {
+    String name =
+        '${PeppermintConstants.adjectives[Random().nextInt(PeppermintConstants.adjectives.length)].capitalizeFirst} ${PeppermintConstants.animals[Random().nextInt(PeppermintConstants.animals.length)].capitalizeFirst}';
+    if (name.length > 32) {
+      name = name.substring(0, 32);
+    }
+
+    return name;
   }
 }