+++ lib/src/widgets/loading.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-
+/// Loading widget
+/// That is use all over the SDK.
 class PeppermintLoading extends StatelessWidget {
   final Color? color;